rename of the CityPlaceCard & CityPlaceList components
