+++ src/components/place-card/place-card.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import CitiesPlaceCard from './cities-place-card.jsx';
+import PlaceCard from './place-card';
 import {offers} from '../../test-data/offers';
 
 Enzyme.configure({
@@ -11,12 +11,12 @@ Enzyme.configure({
 const mockEvent = {
   preventDefault() {},
 };
-describe(`CitiesPlaceCard e2e test`, () => {
-  test(`CitiesPlaceCard should pass offer data to event handler during onPlaceCardMouseEnter event`, () => {
+describe(`PlaceCard e2e test`, () => {
+  test(`PlaceCard should pass offer data to event handler during onPlaceCardMouseEnter event`, () => {
     const onPlaceCardMouseEnter = jest.fn();
 
-    const citiesPlaceCard = shallow(
-        <CitiesPlaceCard
+    const placeCard = shallow(
+        <PlaceCard
           offer={offers[0]}
           onPlaceCardMouseEnter={onPlaceCardMouseEnter}
           index={1}
@@ -24,7 +24,7 @@ describe(`CitiesPlaceCard e2e test`, () => {
         />
     );
 
-    const activeOffer = citiesPlaceCard.find(`.place-card`);
+    const activeOffer = placeCard.find(`.place-card`);
     activeOffer.simulate(`mouseenter`, {});
 
     expect(onPlaceCardMouseEnter.mock.calls[0][0]).toMatchObject(offers[0]);
@@ -34,8 +34,8 @@ describe(`CitiesPlaceCard e2e test`, () => {
 
     const offerIndex = 1;
 
-    const citiesPlaceCard = shallow(
-        <CitiesPlaceCard
+    const placeCard = shallow(
+        <PlaceCard
           offer={offers[0]}
           onPlaceCardMouseEnter={jest.fn()}
           index={offerIndex}
@@ -43,7 +43,7 @@ describe(`CitiesPlaceCard e2e test`, () => {
         />
     );
 
-    const offerTitle = citiesPlaceCard.find(`.place-card__name a`);
+    const offerTitle = placeCard.find(`.place-card__name a`);
     offerTitle.simulate(`click`, mockEvent);
 
     expect(OfferTitleClick.mock.calls.length).toBe(offerIndex);
