+++ src/components/places-list/places-list.test.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import OfferList from './places-list';
+import {offers} from '../../test-data/offers';
+
+test(`OfferList should render correctlt`, () => {
+  const tree = renderer.create(
+      <OfferList offers={offers}
+        onPlaceCardTitleClick={jest.fn()}
+        onPlaceCardMouseEnter={jest.fn()}
+      />).toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
